refactor: move the shared props/icons for dialogs
