+++ src/components/dialogs/common.js
@@ -1,0 +1,15 @@
+export const dialogProps = {
+  kind: {
+    type: String,
+    default: 'info',
+  },
+  header: {
+    type: String,
+    default: 'Here is an informational message.',
+  },
+  message: {
+    type: String,
+    default:
+      'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Duis at tellus at urna condimentum.',
+  },
+};
